meta-mion-simplerunc: Fix Runqueue deadlocking issues
This commit fixes a lot of the Runqueue deadlock slowness issues.

A clean build of host and the two onlp containers now takes around 34
minutes (with no SSTATE and populated DL_DIR) v. 1 hour 53 minutes on
the CI server (with no SSTATE and populated DL_DIR).

There are still more optimizations to be done, but this saves us quite
a bit of time.

Signed-off-by: Eilís Ní Fhlannagáin <pidge@toganlabs.com>