onlpv1_1.0.bb: Complete ONLP install
The original recipe for onlp left a lot of bits out that
are required for platform configuration on onlpdump.

This commits wrangles all those bits and installs them.

A slight change is needed to bsp platform configs mainly
around machines with "-" in them (Standards are great,
everyone has their own. _/-)

TBH this recipe needs some cleaning up after this and
splitting the install up some (we may not want a full
onlp) but this gets us closer to that goal.

Signed-off-by: Eilís Ní Fhlannagáin <pidge@toganlabs.com>